updated score's type to float

+++ models/sentiment.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const sentiment = sequelize.define(
     "sentiment",
     {
-      score: { type: DataTypes.INTEGER, allowNull: false },
+      score: { type: DataTypes.FLOAT, allowNull: false },
       comparativeScore: { type: DataTypes.FLOAT, allowNull: false },
     },
     {}
